Former-commit-id: eee7199f78945369143c4964d41c43fbe5ead4a5

+++ Sources/Tools/AssetCache/ClientNetProxy.cpp
@@ -138,7 +138,9 @@ void ClientNetProxy::OnPacketReceived(DAVA::Net::IChannel* channel, const void* 
     DVASSERT(openedChannel == channel);
     if (length > 0)
     {
-        std::unique_ptr<CachePacket> packet = CachePacket::Create(static_cast<const uint8*>(packetData), length);
+        std::unique_ptr<CachePacket> packet = 
+            CachePacket::Create(static_cast<const uint8*>(packetData), static_cast<uint32>(length));
+
         if (packet != nullptr)
         {
             switch (packet->type)
